07/30/09:
- Did some work on FreeTypeRenderer.hh. Fixed some function definitions 
and prepared to recover my work on the FreeTypeRenderer class. If I dont 
get to it, at least someone will have a better guide to go by.
- Fixed render(...) in FreeTypeRenderer to take a SmartSurface instead 
of an SDL_Surface*. I don't know why it was like that in the first place 
as render(std::string) returned a SmartSurface, so all should have been 
uniform.
- Went through the header files, and got rid of as much of the "using 
std::" as I could. GNU sed is a wonderful tool and I must give it 
credit for doing most of the work.