Did some code formatting.
Player interface redesigned to be consistent with that of the Game classes, to provide cleaner use of iterators and more consistency. CasinoGame and GameBase modified appropriately.

Moves the hand inspection functions from the Player class to the Game class as this is where it belongs. Possibly in the future these functions will be moved to the ruleset classes.